PTE-4: New project structure and pygame window is init in server now
