+++ src/player/assets/characters/cute_boy/idle/0.png
