+++ src/player/assets/characters/cute_boy/idle/1.png
