+++ src/player/assets/characters/cute_boy/run/0.png
