+++ src/player/assets/characters/cute_boy/run/1.png
