+++ src/player/assets/characters/engineer/idle/2.png
