+++ src/player/assets/characters/frog/idle/0.png
