+++ src/player/assets/characters/frog/idle/1.png
