+++ src/player/assets/characters/girl/idle/0.png
