+++ src/player/assets/characters/girl/idle/1.png
