+++ src/player/assets/characters/girl/run/0.png
