+++ src/player/assets/characters/spiderman/idle/0.png
