+++ src/player/assets/characters/spiderman/idle/1.png
