+++ src/player/assets/characters/student/idle/0.png
