+++ src/player/assets/characters/student/idle/1.png
