+++ src/player/assets/characters/student/run/0.png
