+++ src/player/assets/characters/student/run/1.png
